Move base faker to own file
