+++ src/faker.ts
@@ -1,8 +1,7 @@
+import { BaseFaker } from './base-faker';
 import type { LocaleDefinition, MetadataDefinition } from './definitions';
 import { FakerError } from './errors/faker-error';
 import { deprecated } from './internal/deprecated';
-import type { Mersenne } from './internal/mersenne/mersenne';
-import mersenne from './internal/mersenne/mersenne';
 import type { LocaleProxy } from './locale-proxy';
 import { createLocaleProxy } from './locale-proxy';
 import { AirlineModule } from './modules/airline';
@@ -11,214 +10,26 @@ import { ColorModule } from './modules/color';
 import { CommerceModule } from './modules/commerce';
 import { CompanyModule } from './modules/company';
 import { DatabaseModule } from './modules/database';
-import { DatatypeModule } from './modules/datatype';
-import { BaseDateModule, DateModule } from './modules/date';
+import { DateModule } from './modules/date';
 import { FinanceModule } from './modules/finance';
 import { GitModule } from './modules/git';
 import { HackerModule } from './modules/hacker';
-import { BaseHelpersModule, HelpersModule } from './modules/helpers';
+import { HelpersModule } from './modules/helpers';
 import { ImageModule } from './modules/image';
 import { InternetModule } from './modules/internet';
 import type { LocationModule as AddressModule } from './modules/location';
 import { LocationModule } from './modules/location';
 import { LoremModule } from './modules/lorem';
 import { MusicModule } from './modules/music';
-import { NumberModule } from './modules/number';
 import type { PersonModule as NameModule } from './modules/person';
 import { PersonModule } from './modules/person';
 import { PhoneModule } from './modules/phone';
 import { RandomModule } from './modules/random';
 import { ScienceModule } from './modules/science';
-import { StringModule } from './modules/string';
 import { SystemModule } from './modules/system';
 import { VehicleModule } from './modules/vehicle';
 import { WordModule } from './modules/word';
 import { mergeLocales } from './utils/merge-locales';
-
-export class BaseFaker {
-  protected _defaultRefDate: () => Date = () => new Date();
-
-  /**
-   * Gets a new reference date used to generate relative dates.
-   */
-  get defaultRefDate(): () => Date {
-    return this._defaultRefDate;
-  }
-
-  /**
-   * Sets the `refDate` source to use if no `refDate` date is passed to the date methods.
-   *
-   * @param dateOrSource The function or the static value used to generate the `refDate` date instance.
-   * The function must return a new valid `Date` instance for every call.
-   * Defaults to `() => new Date()`.
-   *
-   * @see [Reproducible Results](https://fakerjs.dev/guide/usage.html#reproducible-results)
-   * @see faker.seed() for reproducible results.
-   *
-   * @example
-   * faker.seed(1234);
-   *
-   * // Default behavior
-   * // faker.setDefaultRefDate();
-   * faker.date.past(); // Changes based on the current date/time
-   *
-   * // Use a static ref date
-   * faker.setDefaultRefDate(new Date('2020-01-01'));
-   * faker.date.past(); // Reproducible '2019-07-03T08:27:58.118Z'
-   *
-   * // Use a ref date that changes every time it is used
-   * let clock = new Date("2020-01-01").getTime();
-   * faker.setDefaultRefDate(() => {
-   *   clock += 1000; // +1s
-   *   return new Date(clock);
-   * });
-   *
-   * faker.defaultRefDate() // 2020-01-01T00:00:01Z
-   * faker.defaultRefDate() // 2020-01-01T00:00:02Z
-   */
-  setDefaultRefDate(
-    dateOrSource: string | Date | number | (() => Date) = () => new Date()
-  ): void {
-    if (typeof dateOrSource === 'function') {
-      this._defaultRefDate = dateOrSource;
-    } else {
-      this._defaultRefDate = () => new Date(dateOrSource);
-    }
-  }
-
-  /** @internal */
-  private readonly _mersenne: Mersenne = mersenne();
-
-  readonly datatype: DatatypeModule = new DatatypeModule(this);
-  readonly date: BaseDateModule = new BaseDateModule(this);
-  readonly helpers: BaseHelpersModule = new BaseHelpersModule(this);
-  readonly number: NumberModule = new NumberModule(this);
-  readonly string: StringModule = new StringModule(this);
-
-  /**
-   * Sets the seed or generates a new one.
-   *
-   * Please note that generated values are dependent on both the seed and the
-   * number of calls that have been made since it was set.
-   *
-   * This method is intended to allow for consistent values in tests, so you
-   * might want to use hardcoded values as the seed.
-   *
-   * In addition to that it can be used for creating truly random tests
-   * (by passing no arguments), that still can be reproduced if needed,
-   * by logging the result and explicitly setting it if needed.
-   *
-   * @param seed The seed to use. Defaults to a random number.
-   *
-   * @returns The seed that was set.
-   *
-   * @see [Reproducible Results](https://fakerjs.dev/guide/usage.html#reproducible-results)
-   * @see faker.setDefaultRefDate() when generating relative dates.
-   *
-   * @example
-   * // Consistent values for tests:
-   * faker.seed(42)
-   * faker.number.int(10); // 4
-   * faker.number.int(10); // 8
-   *
-   * faker.seed(42)
-   * faker.number.int(10); // 4
-   * faker.number.int(10); // 8
-   *
-   * // Random but reproducible tests:
-   * // Simply log the seed, and if you need to reproduce it, insert the seed here
-   * console.log('Running test with seed:', faker.seed());
-   */
-  seed(seed?: number): number;
-  /**
-   * Sets the seed array.
-   *
-   * Please note that generated values are dependent on both the seed and the
-   * number of calls that have been made since it was set.
-   *
-   * This method is intended to allow for consistent values in a tests, so you
-   * might want to use hardcoded values as the seed.
-   *
-   * In addition to that it can be used for creating truly random tests
-   * (by passing no arguments), that still can be reproduced if needed,
-   * by logging the result and explicitly setting it if needed.
-   *
-   * @param seedArray The seed array to use.
-   *
-   * @returns The seed array that was set.
-   *
-   * @see [Reproducible Results](https://fakerjs.dev/guide/usage.html#reproducible-results)
-   * @see faker.setDefaultRefDate() when generating relative dates.
-   *
-   * @example
-   * // Consistent values for tests:
-   * faker.seed([42, 13, 17])
-   * faker.number.int(10); // 4
-   * faker.number.int(10); // 8
-   *
-   * faker.seed([42, 13, 17])
-   * faker.number.int(10); // 4
-   * faker.number.int(10); // 8
-   *
-   * // Random but reproducible tests:
-   * // Simply log the seed, and if you need to reproduce it, insert the seed here
-   * console.log('Running test with seed:', faker.seed());
-   */
-  seed(seedArray: number[]): number[];
-  /**
-   * Sets the seed or generates a new one.
-   *
-   * Please note that generated values are dependent on both the seed and the
-   * number of calls that have been made since it was set.
-   *
-   * This method is intended to allow for consistent values in a tests, so you
-   * might want to use hardcoded values as the seed.
-   *
-   * In addition to that it can be used for creating truly random tests
-   * (by passing no arguments), that still can be reproduced if needed,
-   * by logging the result and explicitly setting it if needed.
-   *
-   * @param seed The seed or seed array to use.
-   *
-   * @returns The seed that was set.
-   *
-   * @see [Reproducible Results](https://fakerjs.dev/guide/usage.html#reproducible-results)
-   * @see faker.setDefaultRefDate() when generating relative dates.
-   *
-   * @example
-   * // Consistent values for tests (using a number):
-   * faker.seed(42)
-   * faker.number.int(10); // 4
-   * faker.number.int(10); // 8
-   *
-   * faker.seed(42)
-   * faker.number.int(10); // 4
-   * faker.number.int(10); // 8
-   *
-   * // Consistent values for tests (using an array):
-   * faker.seed([42, 13, 17])
-   * faker.number.int(10); // 4
-   * faker.number.int(10); // 8
-   *
-   * faker.seed([42, 13, 17])
-   * faker.number.int(10); // 4
-   * faker.number.int(10); // 8
-   *
-   * // Random but reproducible tests:
-   * // Simply log the seed, and if you need to reproduce it, insert the seed here
-   * console.log('Running test with seed:', faker.seed());
-   */
-  seed(seed?: number | number[]): number | number[];
-  seed(
-    seed: number | number[] = Math.ceil(Math.random() * Number.MAX_SAFE_INTEGER)
-  ): number | number[] {
-    this._mersenne.seed(seed);
-
-    return seed;
-  }
-}
-
-export const baseFaker = new BaseFaker();
 
 /**
  * This is Faker's main class containing all modules that can be used to generate data.
